appointment-service/requirements.txt
aerich==0.9.1
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
asyncclick==8.1.8.0
dictdiffer==0.9.0
idna==3.10
iso8601==2.1.0
pydantic==2.11.7
pydantic_core==2.33.2
pypika-tortoise==0.6.1
pytz==2025.2
sniffio==1.3.1
tortoise-orm==0.25.1
typing-inspection==0.4.1
typing_extensions==4.14.1
grpcio
grpcio-tools
redis>=4.2